separate app.js file for lib imports and middlewares v1.2
